# entityBackend/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Issuer & Service Provider backend
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /entity
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - text/plain
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/plain
paths:
...
  /presentation:
    # binds a127 app logic to a route
    x-swagger-router-controller: entity.controller
    post:
      description: Get data from a presentation
      # used as the method name of the controller
      operationId: recivePresentationData
      parameters:
        - name: presentation
          in: body
          description: Necessary data to create a new Alastria ID
          required: false
          schema: 
            $ref: "#/definitions/Presentation"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/PresentationData"
        # responses may fall through to errors
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Address not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /alastriaToken:
    # binds a127 app logic to a route
    x-swagger-router-controller: entity.controller
    post:
      description: Verify Alastria Session from Wallet
      # used as the method name of the controller
      operationId: verifyAlastriaSession
      parameters:
        - name: alastriaSession
          in: body
          description: JWT signed by Wallet to create an Alastria Session
          required: true
          schema: 
            $ref: "#/definitions/identityData"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/userTokenAccess"
        # responses may fall through to errors
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Address not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /alastria/identity:
    # binds a127 app logic to a route
    x-swagger-router-controller: entity.controller
    post:
      description: Create a new Alastria identity, sent to an entity acting as an issuer (or identity helper)
      # used as the method name of the controller
      operationId: createAlastriaID
      parameters:
        - name: body
          in: body
          description: Necessary data to create a new Alastria ID
          required: false
          schema: 
            $ref: "#/definitions/identityData"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/alastriaIdentityResponse"
        # responses may fall through to errors
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Address not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /alastria/identity/{alastriaDID}:
    # binds a127 app logic to a route
    x-swagger-router-controller: entity.controller
    get:
      description: Get public key of a given Alastria DID
      operationId: getCurrentPublicKey
      parameters:
        - name: alastriaDID
          in: path
          description: AlastriaId to find the publicKey
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/publickeyResponse"
        # responses may fall through to errors
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Address not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /alastria/credential:
    # binds a127 app logic to a route
    x-swagger-router-controller: entity.controller
    get:
      description: Get available credentials catalog from an issuer
      operationId: getCredentialCatalog
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/credentialCatalogList"
        # responses may fall through to errors
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Address not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      # The user type (subject or Issuer) could be determined from the entity session -> DID user type
      # for the MVP we could assume the call is from an issuer the implementation for a subject will be required in the future
      description: Registry Credential by the issuer
      # used as the method name of the controller
      operationId: addIssuerCredential
      parameters:
        - name: credential
          in: body
          description: Necessary data to add a new subject credential
          required: false
          schema: 
            $ref: "#/definitions/issuerCredentialdata"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/addCredentialResponse"
        # responses may fall through to errors
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Address not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /alastria/credential/status/{credentialHash}:
    get:
      x-swagger-router-controller: entity.controller
      description: get current Issuer or Subject status for the given credential
      operationId: getCredentialStatus
      parameters:
        - name: credentialHash
          in: path
          type: string
          description: credential hash
          required: true
        # only one of issuer & subject parameters should be used to obtain the Issuer status or the Subject status 
        - name: issuer
          in: query
          type: string
          description: isssuer did
          required: false
        - name: subject
          in: query
          type: string
          description: subject did
          required: false
      responses: 
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/addCredentialResponse"
        # responses may fall through to errors
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Address not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      # for the MVP we only need post from the issuer but a post for the subject to delete the credential will be required in the future 
      description: Update an issuer credential status if posible
      # used as the method name of the controller
      operationId: updateCredentialStatus
      parameters:
        - name: credentialHash
          in: path
          type: string
          description: credential hash
          required: true
        - name: body
          in: body
          description: Necessary data to add change credential status
          required: true
          schema: 
            $ref: "#/definitions/updateCredentialData"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/addCredentialResponse"
        # responses may fall through to errors
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Credential not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /alastria/presentation/status/{presentationHash}:
    get:
      x-swagger-router-controller: entity.controller
      description: get current Subject or Service Provider (Receiver) status for the given presentation
      operationId: getPresentationStatus
      parameters:
        - name: presentationHash
          in: path
          type: string
          description: presentation hash
          required: true
        # only one of serviceProvider & subject parameters should be used to obtain the Issuer status or the Subject status 
        - name: serviceProvider
          in: query
          type: string
          description: serviceProvider did
          required: false
        - name: subject
          in: query
          type: string
          description: subject did
          required: false
      responses: 
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/addPresentationResponse"
        # responses may fall through to errors
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Address not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      # The user type (subject or ServiceProvider) could be determined from the entity session -> DID user type
      # for the MVP we could assume the call is from a ServiceProvider, the implementation for a subject will be required in the future    
      x-swagger-router-controller: entity.controller
      description: Update a serviceProvider presentation status if posible
      # used as the method name of the controller
      operationId: updateReceiverPresentationStatus
      parameters:
        - name: presentationHash
          in: path
          type: string
          description: presentation hash
          required: true
        - name: body
          in: body
          description: Necessary data to add change presentation status
          required: true
          schema: 
            $ref: "#/definitions/updatePresentationData"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/addPresentationResponse"
        # responses may fall through to errors
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: presentation not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  userTokenAccess:
    properties:
      authtoken:
        type: object
...
  PresentationData:
    properties:
      data:
        type: string
  Presentation:
    properties:
      signedPresentation:
        type: object
  # credential:
  #   properties:
  #     header:
  #       type: "object"
  #     payload:
  #       type: "object"
  #     signature:
  #       type: "object"
  alastriaIdentityResponse:
    properties:
      alastriaDID:
        type: string
  credentialCatalogList:
    properties: 
      list:
        type: string
        #schema:
          #$ref: "#/definitions/credentialCatalog"
  # credentialCatalog:
  #   properties:
  #     credentialName:
  #       type: string
  #     status:
  #       type: string
  #     credentialType:
  #       type: string
  addCredentialResponse:
    properties:
      status:
        type: string
  updateCredentialData:
    properties:
      newStatus:
        type: string
  publickeyResponse:
    properties:
      publicKey:
        type: string
...
  addPresentationResponse:
    properties:
      status:
        type: string
  updatePresentationData:
    properties:
      newStatus:
        type: number
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  identityData:
    type: "object"
    required: [signedAIC]
    properties:
      "signedAIC":
        type: string
  issuerCredentialdata:
    type: "object"
    required: [issuerCredential]
    properties:
      "issuerCredential":
        type: object
...